Implemented BLE duplicate packet filtering and baseline logic across Distance Master, Ladder Drill, and Club Gapping features. Added game lifecycle management events to ensure proper resource cleanup and sync timer resets.
- Added baseline packet handling and duplicate filtering to prevent redundant shot recording in `DistanceMasterBloc`, `LadderDrillBloc`, and `ClubGappingBloc`.
- Updated `DistanceMasterBloc` and `LadderDrillBloc` to handle carry distance as `double` instead of `int` for higher precision.
- Introduced `EndGameEvent` and `RestartLadderDrillGameEvent` to properly cancel BLE subscriptions and reset game states.
- Implemented club-specific command packets and baseline resets in `ClubGappingBloc` when switching or retaking clubs.
- Updated `main_screen.dart` to trigger cleanup events when navigating between features.
- Fixed navigation routes in `sign_in_screen.dart` to point to the main screen.
- UI refinement: Updated carry distance display to show 1 decimal point in `shot_recording_page.dart`.

+++ lib/feature/distance_control_drills/ladder_drill/presentation/bloc/ladder_drill_state.dart
@@ -75,9 +75,9 @@ class GameInProgressState extends LadderDrillState {
   final int streak;
   final int maxStreak;
 
-  int? get lastActualCarry {
+  double? get lastActualCarry {
     if (currentShots.isEmpty) return null;
-    return currentShots.last.carryDistance.toInt();
+    return currentShots.last.carryDistance;
   }
 
   bool get isLastShotWithinTarget {
